feat: Refactor Header and Footer components to utilize data props and streamline layout queries
